interface for middlewareBuilder
